+++ app/src/main/java/com/cswala/cswala/Activities/LoginActivity.java
@@ -15,6 +15,7 @@ import android.widget.Toast;
 import androidx.annotation.NonNull;
 import androidx.appcompat.app.AppCompatActivity;
 
+import com.cswala.cswala.MainActivity;
 import com.cswala.cswala.R;
 import com.cswala.cswala.utils.IntentHelper;
 import com.google.android.gms.auth.api.signin.GoogleSignIn;
