+++ app/src/main/java/com/cswala/cswala/MainActivity.java
@@ -1,4 +1,4 @@
-package com.cswala.cswala.Activities;
+package com.cswala.cswala;
 
 import android.content.Intent;
 import android.os.Bundle;
@@ -8,12 +8,12 @@ import android.widget.TextView;
 import androidx.appcompat.app.AppCompatActivity;
 import androidx.fragment.app.Fragment;
 
+import com.cswala.cswala.Activities.LoginActivity;
 import com.cswala.cswala.Fragments.CommunityFragment;
 import com.cswala.cswala.Fragments.ExploreFragment;
 import com.cswala.cswala.Fragments.JobHunt;
 import com.cswala.cswala.Fragments.NewsFragment;
 import com.cswala.cswala.Fragments.ProfileFragment;
-import com.cswala.cswala.R;
 import com.google.firebase.auth.FirebaseAuth;
 import com.ismaeldivita.chipnavigation.ChipNavigationBar;
