+++ app/src/main/java/com/cswala/cswala/utils/IntentHelper.java
@@ -6,7 +6,7 @@ import android.content.Intent;
 import com.blogspot.atifsoftwares.animatoolib.Animatoo;
 import com.cswala.cswala.Activities.IntroClass;
 import com.cswala.cswala.Activities.LoginActivity;
-import com.cswala.cswala.Activities.MainActivity;
+import com.cswala.cswala.MainActivity;
 
 public class IntentHelper {
     Context context;
